feat: move testing utilities to new directory
