+++ test/test_ca.py
@@ -1,7 +1,7 @@
 import time
 
 import j1939
-from feeder import Feeder
+from test_helpers.feeder import Feeder
 
 def address_claim(
     feeder,
