+++ test/test_ecu.py
@@ -1,7 +1,7 @@
 import time
 
 import j1939
-from feeder import Feeder
+from test_helpers.feeder import Feeder
